Implement a translation build
Include a rollup config and command to directly create an ESM version of
the sqlite.js file directly from sources that would otherwise be used
by webpack to build the other version, since this thing uses an old and
decrepit version of webpack that does not seem to be capable of
generating an ESM file on its own.

This file also bundles within it the big-integer package, which is the
only dependency this has, which means that this file is a single file
dependency now (which is nice cause SQLite as a C library does the
same thing).

There is also added an additional test file that exercises this new
version using the existing SQLite tests, to ensure that the built
version is fully functional.