Do some package cleanup
This removes a bunch of stuff from the repo that we do not want, or will
want to replace with something else, in order to get things as small and
lean as is currently possible.

This swaps us to pnpm, remotes eslint, and the githooks that we do not
want, along with some CI config for things we are not currently using.

Also, not using editorconfig and not using vscode because bleh.