Whack everything but SQLite
This removes all of the other dialects except SQLite and adjusts tests
as appropriate to get as many of them running as possible. Most/all of
the tests here are failing because the SQLite parser does not accept
them, which is  good enough for the moment while trying to get thing
going.

Since the tests are a fundamental part of the build, the build also had
to be adjusted, to remove the code that tries to generate type
bindings, which I am not interested in.

Some of this is still a bit crude, in as much as various tests were
trying to us specific dialects and they're all forced into using
SQLite.

A later step will adjust the API to not take this option and just always
use the only parser.