Change the API spec for single dialect
Previously there were multiple dialects, with one of them being a
default, and you could specify a dialect as a database option in the
options to the parser. Since there is only a single dialect now, this
aspect has been removed.

In order to test this, the parser currently complains and throws an
error if you specify this field in the options, so that all places can
be ferreted out.

This change also touches some places where the database was actually
used previously in a way that makes me nervous I didn't notice it
sooner, and also ensures that all of the tests no longer specify one,
although this could be cleaner.