Replace the inline peggy plugin
In order to make the grammar a bit easier to work with since the file is
currently massive, this implements a new rollup plugin dedicated to
peggy.js compilations, in a separate file in order to be reusable in
other projects as needed.

This plugin allows for mapping a virtual filename to a configuraton
object, so that a grammar can be pulled into separate parts and still
be compiled together appropriately.

This is currently a bit overkill since it transpires that every file
must have at least one rule in it, and they are all combined together
regardless of their order, based on the start token. In suh a case the
plugin could just pull all files in a folder and combine them, but that
realization came too late, and this provides more power.